Refactor format factories
