+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/factories/TableFactory.scala
@@ -21,7 +21,14 @@ package org.apache.flink.table.factories
 import java.util
 
 /**
-  * Common trait for all properties-based discoverable table factories.
+  * A factory to create different table-related instances from string-based properties. This
+  * factory is used with Java's Service Provider Interfaces (SPI) for discovering. A factory is
+  * called with a set of normalized properties that describe the desired format. The factory allows
+  * for matching to the given set of properties.
+  *
+  * Classes that implement this interface can be added to the
+  * "META_INF/services/org.apache.flink.table.factories.TableFactory" file of a JAR file in
+  * the current classpath to be found.
   */
 trait TableFactory {
